requirements.txt
python-dotenv
mysql-connector-python
pandas
scikit-learn
matplotlib